+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+-   `intent` property to `Button`, `FieldItem`, `FieldSet`, `Notification`, and `Spinner`
+
 ## [24.1.0] - 2025-04-16
 
 ### Added